[#171994544] Refactoring
Refactoring
